Post DC commit
- Seven shot autonomous
   - not final, I'd like to work on an encoder-gyro control
     system.
- No new functionality, only bug fixes.
- This week I'm going to:
   - figure out why our cRio stopped working in finals.
   - get rid of smartdashboard, it sucks.
   - clean up code.

+++ src/WiredCatsEvents/SensorEvents/EventPositionChanged.java
@@ -12,11 +12,11 @@ import WiredCatsEvents.WiredCatsEvent;
  */
 public class EventPositionChanged extends WiredCatsEvent
 {
-    public int leftEncoder;
-    public int rightEncoder;
+    public double leftEncoder;
+    public double rightEncoder;
     public double time;
     
-    public EventPositionChanged(Object source, int leftEncoder, int rightEncoder, double time)
+    public EventPositionChanged(Object source, double leftEncoder, double rightEncoder, double time)
     {
         super(source);
         this.leftEncoder = leftEncoder;
